requirements.txt
pandas
openai
python-dotenv
ipykernel
